Moved message's related classes to different package
