+++ app/src/main/java/com/teamagam/gimelgimel/app/injectors/modules/RepositoryModule.java
@@ -16,10 +16,10 @@ import com.teamagam.gimelgimel.data.map.repository.SingleDisplayedItemDataReposi
 import com.teamagam.gimelgimel.data.notifications.PersistentConnectivityStatusRepositoryImpl;
 import com.teamagam.gimelgimel.data.rasters.repository.IntermediateRasterVisibilityDataRepository;
 import com.teamagam.gimelgimel.data.rasters.repository.IntermediateRastersRepositoryData;
-import com.teamagam.gimelgimel.data.response.repository.MessagesDataRepository;
-import com.teamagam.gimelgimel.data.response.repository.NewMessageIndicationDataRepository;
-import com.teamagam.gimelgimel.data.response.repository.ObjectMessageDataMapper;
-import com.teamagam.gimelgimel.data.response.repository.UnreadMessagesCountDataRepository;
+import com.teamagam.gimelgimel.data.message.repository.MessagesDataRepository;
+import com.teamagam.gimelgimel.data.message.repository.NewMessageIndicationDataRepository;
+import com.teamagam.gimelgimel.data.message.repository.ObjectMessageDataMapper;
+import com.teamagam.gimelgimel.data.message.repository.UnreadMessagesCountDataRepository;
 import com.teamagam.gimelgimel.data.sensors.repository.SelectedSensorDataRepository;
 import com.teamagam.gimelgimel.data.sensors.repository.SensorsDataRepository;
 import com.teamagam.gimelgimel.data.user.repository.UserPreferenceRepositoryImpl;
