+++ data/src/main/java/com/teamagam/gimelgimel/data/message/repository/InMemory/InMemoryMessagesCache.java
@@ -1,4 +1,4 @@
-package com.teamagam.gimelgimel.data.response.repository.InMemory;
+package com.teamagam.gimelgimel.data.message.repository.InMemory;
 
 import com.teamagam.gimelgimel.data.base.repository.SubjectRepository;
 import com.teamagam.gimelgimel.domain.messages.entity.ChatMessage;
