+++ data/src/main/java/com/teamagam/gimelgimel/data/message/repository/NewMessageIndicationDataRepository.java
@@ -1,4 +1,4 @@
-package com.teamagam.gimelgimel.data.response.repository;
+package com.teamagam.gimelgimel.data.message.repository;
 
 import com.teamagam.gimelgimel.domain.messages.repository.NewMessageIndicationRepository;
 import java.util.Date;
