+++ data/src/main/java/com/teamagam/gimelgimel/data/message/repository/cache/room/Converter.java
@@ -1,0 +1,16 @@
+package com.teamagam.gimelgimel.data.message.repository.cache.room;
+
+import android.arch.persistence.room.TypeConverter;
+import java.util.Date;
+
+public class Converter {
+  @TypeConverter
+  public static Date fromTimestamp(Long value) {
+    return value == null ? null : new Date(value);
+  }
+
+  @TypeConverter
+  public static Long dateToTimestamp(Date date) {
+    return date == null ? null : date.getTime();
+  }
+}
